Update parmap API to remove DeprecationWarnings
parmap 1.5.0 introduced support for passing keyword arguments to
mapped functions. The drawback of this feature is that a consistent
naming of parmap arguments is needed in order to prevent conflicts
between parmap arguments and user-defined arguments. The new parmap
convention is to prefix all its arguments with "pm_" (for parmap)
as documented http://parmap.readthedocs.io/en/latest/

This commit updates the sifra parmap calls to use the new API,
removing the deprecation warnings that appear with
parmap==1.5.1 (the parmap version used).

It also upgrades the requirements.txt file, as parmap >= 1.5.0
is now required.